+++ server/iocContainer/index.ts
@@ -1499,6 +1499,7 @@ export default async function getBottle(
       async (params) =>
         container.NcmecService.getUserHasExistingNcmecReport(params),
       container.ManualReviewContentResolver,
+      container.Meter,
     );
   });